Add rpc_decodescript.py test
