+++ test/functional/test_runner.py
@@ -100,7 +100,7 @@ BASE_SCRIPTS= [
 #    'feature_proxy.py',
 #    'rpc_signrawtransaction.py',
 #    'p2p_disconnect_ban.py',
-#    'rpc_decodescript.py',
+   'rpc_decodescript.py',
    'rpc_blockchain.py',
 #    'rpc_deprecated.py',
 #    'wallet_disable.py',
